requirements.txt
# PIP
invoke==1.0.0
pip==10.0.1
prompt_toolkit==1.0.15
pytest==3.6.1
six==1.11.0
wcwidth==0.1.7

# OE Stack
OpenEye-Artemis==0.2.1a5
OpenEye-cuberecord==0.14.7
OpenEye-datarecord==0.12.2
OpenEye-ensemble2img==0.1.1
OpenEye-floe==0.5.18
OpenEye_oetrajanalysis==0.1.3
OpenEye-orionclient==0.5.1
openeye-toolkits==2018.2.1



